Fix svelte warning and my own gut warning
Probably the register should only happen once, no matter how many components are rendered in the app.

Also the `swiperEl` was apparently not... updating...? Seemed fine to me. Perhaps the problem was with derivations and effects not being properly triggered (possibly?). I think this isn't a problem here, but Svelte isn't aware. All it sees is that the element in question is behind an `if` block, but that the evaluation of that block is always true in the browser.